Latest generated openwire
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/trunk@384828 13f79535-47bb-0310-9956-ffa450edef68

+++ openwire-cpp/src/marshal/ResponseMarshaller.cpp
@@ -56,7 +56,7 @@ void ResponseMarshaller::unmarshal(ProtocolFormat& wireFormat, Object o, BinaryR
     base.unmarshal(wireFormat, o, dataIn, bs);
 
     Response& info = (Response&) o;
-        info.setCorrelationId(dataIn.readShort());
+        info.setCorrelationId(dataIn.readInt());
 
 }
 
@@ -79,6 +79,6 @@ void ResponseMarshaller::marshal2(ProtocolFormat& wireFormat, Object& o, BinaryW
     base.marshal2(wireFormat, o, dataOut, bs);
 
     Response& info = (Response&) o;
-    DataStreamMarshaller.writeShort(info.getCorrelationId(), dataOut);
+    DataStreamMarshaller.writeInt(info.getCorrelationId(), dataOut);
 
 }
